Update stack.c
Hey guys,

So here's the problem. When we malloc a bunch of memory and then fill it the programme works fine. The issues start if we try to realloc to a smaller block of memory. Example is that when we first create and populate memory with say 10 ints and then try to realloc to a size of 9 ints we automatically fail because 9 is not large enough to hold ten. I'm trying to write an if statement which will check for this and fail the 'setStackSize' function if the user attempts to resize their memory down to something too small. 

Basically our malloc, realloc and delete functions all work fine. We just have to make sure Bodo doesn't set the size too small and crashes the programme. 

So two final things to do -  make the inifinite stack and program that if statement to check that the new realloced memory is big enough for the current population of stored values. Will continue with this later today and hopefully have it solved. Will post if I get anywhere.