Update threads.c
So to solve the problem of the program just skipping over the first message of the parent I've added a mutex around this to. Basically this locks the program into completing in the right order. The sleep command is used to force the OS to reschedule to thread 2. I noticed that it was getting stuck in thread 1 waiting for the signal command. By forcing it to switch it works perfectly every time.